+++ src/main/java/com/bancoeconomico/service/Saldo.java
@@ -6,6 +6,6 @@ import java.math.BigDecimal;
 
 public interface Saldo<T extends Cliente> {
 
-    BigDecimal consultarSaldo(T cliente, Integer numeroConta);
+    BigDecimal consultarSaldo(T cliente, Long numeroConta);
 
 }
