+++ src/main/java/com/bancoeconomico/service/Transferencia.java
@@ -7,6 +7,6 @@ import java.math.BigDecimal;
 
 public interface Transferencia<T extends Cliente> {
 
-    void transferir(T cliente, Integer numeroContaOrigem, Conta destino, BigDecimal valor);
+    void transferir(T cliente, Long numeroContaOrigem, Conta destino, BigDecimal valor);
 
 }
